% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_disableActionButton.R
\name{utils_disableActionButton}
\alias{utils_disableActionButton}
\title{Disable Action Button}
\usage{
utils_disableActionButton(id, session)
}
\arguments{
\item{val1}{href domain}

\item{val2}{text}
}
\value{
string
}
\description{
Yup
}
